+++ src/scg_exact_scg.c
@@ -48,7 +48,7 @@ int igraph_i_exact_coarse_graining(const igraph_real_t *v,
     w[i].ind = i;
   }
 
-  qsort(w, n, sizeof(igraph_i_scg_indval_t), igraph_i_compare_ind_val);
+  qsort(w, (size_t) n, sizeof(igraph_i_scg_indval_t), igraph_i_compare_ind_val);
   
   gr_nb = 0;
   gr[w[0].ind] = gr_nb;
